Color Icon working
Hooked the secure function to be able to change the border color of a
needed item when the loot table is open. Will adjust color to make it as
noticeable as possible. Also moved sound so that it will only go off
once if multiple needed items are looted at once